// mobile/src/lib/types.ts
export interface Product {
  id: string;
  codigoProduto: string;
  descricaoProduto: string;
  status: boolean;
  fotoProduto: string | null;
  thumbnailUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PaginatedProducts {
  data: Product[];
  meta: {
    total: number;
    page: number;
    limit: number;
    totalPages: number;
  };
}

export interface User {
  id: string;
  name: string;
  email: string;
}

export interface AuthResponse {
  accessToken: string;
  user: User;
}